% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_rbind.R
\name{better.rbind}
\alias{better.rbind}
\title{Bind Lists Of Data Frames By Row}
\usage{
better.rbind(listlist.tables, sort.byrow = NULL, ...)
}
\arguments{
\item{listlist.tables}{The list of list of tables that will be row binded into a single list of tables.}

\item{sort.byrow}{The index of the row that will be used to sort the whole table by. If NULL the tables won't be sorted.}

\item{...}{Parameters that will be passed on to the order function. order is used for sorting the tables.}
}
\value{
A list of the row binded data frames.
}
\description{
This function takes a list of lists of data frames as argument. The function row binds the first tables of every sublist, the second tables of every
sublist, the third tables of every sublist and so on together. The result is a single list of tables. It is also possible to sort the tables by any
row after row binding them.
}
