% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_attachment_elabftw.R
\name{check.attachment.elabftw}
\alias{check.attachment.elabftw}
\title{Check The Attached Files Of A Entry On eLabFTW}
\usage{
check.attachment.elabftw(expID, extension = c(".csv", ".txt"))
}
\arguments{
\item{expID}{The unique id of the eLabFTW entry. The id can be found in the url of the labbook entry.}

\item{extension}{The file extension the function recognises. Only strings and factors ending in one of the given extensions will be counted as missing
file, if they do not match any of the attached file names.}
}
\value{
A list containing three vectors: A vector of all file names that are attached to the labbook entry but never mentioned in any of the tables,
a vector of file names that are mentioned in the tables of the labbook entry, but are not attached to the labbook entry, and a vector of all file names
that are attached multiple times.
}
\description{
This function downloads all tables of a labbook entry in the online labbook eLabFTW together with a list of all files that are attached to the entry.
The list of file names will be compared to all table entries. All file names that are not mentioned in the tables will be returned. All file names
that are mentioned in the tables but not attached will be returned. All file names of files that are attached multiple times will be returned. eLabFTW
allows the user to attach different files with the exact same file name for some fucking reason. The purpose of this function is spell checking the
file names and finding missing attachments to ensure a smooth processing of the tables.
}
