% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_version.R
\name{check.version}
\alias{check.version}
\title{Check The Version Of The Package}
\usage{
check.version(expected.version, package = "RHotStuff", strict = FALSE, ...)
}
\arguments{
\item{expected.version}{The version (as string) of the package the script was designed to use.}

\item{package}{The package that will be checked. A vector of packages can't be handled and will
be ignored; except the first element of the vector!}

\item{strict}{If set TRUE, the expected version must match exactly the version of the package.
If set FALSE, not matching, but compatible versions will also be excepted.}

\item{...}{Parameters passed on to packageVersion() function.}
}
\value{
No return value. The function throws a warning if necessary.
}
\description{
This function makes sure that the scripts your running is working with the correct version
of RHotStuff. If the version does not match the expectations, the function prompts a warning.
}
\details{
This function can check also other packages, but it does not except a vector of package names.
The versions must be a string in the format "x.y.z". If the parameter strict is set FALSE, x
must match exactly, but two version will be considered equal if y and z are just greater or
equal the expected version.
}
