% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_elabftw.R
\name{GET.elabftw.bycaption}
\alias{GET.elabftw.bycaption}
\title{Request Experimental Data From eLabFTW Via API}
\usage{
GET.elabftw.bycaption(
  experiment.id,
  caption = "Messdaten",
  dec = ",",
  header = F,
  api.key = Sys.getenv("READ_ELABFTW_TOKEN"),
  url = Sys.getenv("ELABFTW_API_URL"),
  outputHTTP = F
)
}
\arguments{
\item{experiment.id}{The eLabFTW id of the experiment that will be fetched. The id can be found in the url of the experiment.}

\item{caption}{A string identifying the table. The caption must be placed in the first cell of the first row of the table.}

\item{dec}{The decimal separator used in the table. The default is the German number notation with a comma instead of a period!}

\item{header}{Is the first line of the table the column names or data? If true, the first row is no data.}

\item{api.key}{The eLabFT API-key. See the eLabFTW documentation for details. The default settings retrieve this value from
the enviourmental variable READ_ELABFTW_TOKEN. This variable can be set in the ~/.Renviron file.}

\item{url}{The url to the eLabFTW API. See the eLabFTW documentation for details. The default settings retrieve this value from
the enviourmental variable ELABFTW_API_URL. This variable can be set in the ~/.Renviron file.}

\item{outputHTTP}{Boolean value. If FALSE only the selected tables  will be returned. If TRUE the selected tables and the original http response received from the API will
be returned. The http response contains meta information needed by other parsing functions.}
}
\value{
A list of data frames with the content of the selected html tables from the eLabFTW online labbook. If outputHTTP is
set TRUE the return value is a list containing the list of all data frames created from the selected html tables and the
original http response received from the API.
}
\description{
This function fetches experimental data from the online labbook eLabFTW as a data frame. The html table that will be returned
as data frame must be marked with a caption. The first cell of the first row of the table will be interpreted as the caption.
The caption is removed from the final data frame.
}
\details{
Keep in mind: Tables that have a caption can't be parsed by GET.elabftw.byselector().
}
