% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hot.R
\name{theme_hot}
\alias{theme_hot}
\title{Custom Theme For GGPlot2}
\usage{
theme_hot()
}
\value{
No idea what funky shit ggplot2 does. Use the return value like any other ggplot2 theme.
}
\description{
This function wraps a bunch of ggplot2 commands to style the plots in a consistent, easy to code and pretty hot way.
}
